Removed base_sql.py
